feat: add auth handler
